Add failAfterTime to limit maximum time elapsed
This commit adds the failAfterTime method to Backoff (and FunctionCall)
to allow limiting the total elapsed time.  The primary intended use case
is when the caller has a deadline which can not be exceeded during which
to retry an operation with backoff and the number of backoffs can not be
calculated in advance (e.g. because the operation takes a variable
amount of time).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>